requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
cffi==2.0.0
Django==6.0.3
psycopg2-binary==2.9.11
pycparser==3.0
sqlparse==0.5.5
tzdata==2025.3
